% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info_covid19tunisia}
\alias{get_info_covid19tunisia}
\title{Get information about the datasets provided by covid19tunisia}
\usage{
get_info_covid19tunisia()
}
\value{
A tibble with information about where the data is pulled from, details about the dataset, what the data types are, etc.
}
\description{
Get information about the datasets provided by covid19tunisia
}
\examples{
get_info_covid19tunisia()
}
