% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh_covid19tunisia}
\alias{refresh_covid19tunisia}
\title{Refresh data}
\usage{
refresh_covid19tunisia(verbose = TRUE)
}
\arguments{
\item{verbose}{Should messages about what's going on be displayed?}
}
\value{
A tibble of the most up to date per-day data on cases of COVID-19 in Tunisia, ordered by date descending.
}
\description{
Refresh data
}
\examples{
\donttest{
refresh_covid19tunisia()
}
}
